refactor(ecstore): move List/Bucket Operations into set_disk::ops (backlog#819)
P3+P4 of the SetDisks God-Object split (tracking backlog#815, depends on P0
backlog#816).

Give the List and Bucket operation families their own module homes:

- set_disk/ops/bucket.rs: relocate `impl BucketOperations for SetDisks`
  (make_bucket / get_bucket_info / list_bucket / delete_bucket) out of the
  set_disk/mod.rs God-Object.
- set_disk/ops/list.rs: consolidate the `impl ListOperations for SetDisks`
  storage-api contract impl (relocated from mod.rs) with the borrow-based
  `SetDisks::delete_all` service unit and its `ListOperations<'a>` handle
  (relocated from set_disk/list.rs, which is removed).
- Repoint the one external reference to the service-unit struct in
  set_disk/lock.rs tests to crate::set_disk::ops::list::ListOperations.

Both contracts stay implemented for SetDisks so their associated-type bounds
are unchanged. Method bodies are moved verbatim — sorted content comparisons
against the pre-move sources are byte-identical modulo the `use super::*;` →
`use super::super::*;` header swap. Runtime behavior is unchanged.
