+++ crates/ecstore/src/set_disk/mod.rs
@@ -444,7 +444,6 @@ const DEFAULT_RUSTFS_GET_METADATA_VERSION_EARLY_STOP_ENABLE: bool = false;
 static OBJECT_LOCK_DIAG_ENABLED: OnceLock<bool> = OnceLock::new();
 
 mod ctx;
-mod list;
 mod lock;
 mod metadata;
 mod ops;
@@ -3532,213 +3531,6 @@ impl crate::storage_api_contracts::namespace::NamespaceLocking for SetDisks {
     }
 }
 
-#[async_trait::async_trait]
-impl BucketOperations for SetDisks {
-    type Error = Error;
-
-    #[tracing::instrument(skip(self))]
-    async fn make_bucket(&self, bucket: &str, opts: &MakeBucketOptions) -> Result<()> {
-        let disks = self.disk_inventory().await;
-        let write_quorum = (disks.len() / 2) + 1;
-        let force_create = opts.force_create;
-
-        let mut futures = Vec::with_capacity(disks.len());
-        for disk in disks {
-            let bucket = bucket.to_string();
-            futures.push(async move {
-                match disk {
-                    Some(disk) => match disk.make_volume(&bucket).await {
-                        Ok(()) => Ok(()),
-                        Err(err) if force_create && matches!(err, DiskError::VolumeExists) => Ok(()),
-                        Err(err) => Err(err),
-                    },
-                    None => Err(DiskError::DiskNotFound),
-                }
-            });
-        }
-
-        let results = join_all(futures).await;
-        let errs = results
-            .into_iter()
-            .map(|result| result.err())
-            .collect::<Vec<Option<DiskError>>>();
-
-        if let Some(err) = reduce_write_quorum_errs(&errs, BUCKET_OP_IGNORED_ERRS, write_quorum) {
-            return Err(err.into());
-        }
-
-        Ok(())
-    }
-
-    #[tracing::instrument(skip(self))]
-    async fn get_bucket_info(&self, bucket: &str, _opts: &BucketOptions) -> Result<BucketInfo> {
-        let disks = self.disk_inventory().await;
-        let write_quorum = (disks.len() / 2) + 1;
-
-        let mut futures = Vec::with_capacity(disks.len());
-        for disk in disks {
-            let bucket = bucket.to_string();
-            futures.push(async move {
-                match disk {
-                    Some(disk) => disk.stat_volume(&bucket).await,
-                    None => Err(DiskError::DiskNotFound),
-                }
-            });
-        }
-
-        let results = join_all(futures).await;
-        let mut infos = Vec::with_capacity(results.len());
-        let mut errs = Vec::with_capacity(results.len());
-        for result in results {
-            match result {
-                Ok(info) => {
-                    infos.push(Some(info));
-                    errs.push(None);
-                }
-                Err(err) => {
-                    infos.push(None);
-                    errs.push(Some(err));
-                }
-            }
-        }
-
-        if let Some(err) = reduce_write_quorum_errs(&errs, BUCKET_OP_IGNORED_ERRS, write_quorum) {
-            return Err(err.into());
-        }
-
-        let mut versioning = false;
-        let mut object_locking = false;
-        if let Ok(sys) = metadata_sys::get(bucket).await {
-            versioning = sys.versioning();
-            object_locking = sys.object_locking();
-        }
-
-        infos
-            .into_iter()
-            .flatten()
-            .next()
-            .map(|info| BucketInfo {
-                name: info.name,
-                created: info.created,
-                versioning,
-                object_locking,
-                ..Default::default()
-            })
-            .ok_or(Error::VolumeNotFound)
-    }
-
-    #[tracing::instrument(skip(self))]
-    async fn list_bucket(&self, _opts: &BucketOptions) -> Result<Vec<BucketInfo>> {
-        let disks = self.disk_inventory().await;
-        let write_quorum = (disks.len() / 2) + 1;
-
-        let mut futures = Vec::with_capacity(disks.len());
-        for disk in disks {
-            futures.push(async move {
-                match disk {
-                    Some(disk) => disk.list_volumes().await,
-                    None => Err(DiskError::DiskNotFound),
-                }
-            });
-        }
-
-        let results = join_all(futures).await;
-        let mut infos = Vec::with_capacity(results.len());
-        let mut errs = Vec::with_capacity(results.len());
-        for result in results {
-            match result {
-                Ok(volumes) => {
-                    infos.push(Some(volumes));
-                    errs.push(None);
-                }
-                Err(err) => {
-                    infos.push(None);
-                    errs.push(Some(err));
-                }
-            }
-        }
-
-        if let Some(err) = reduce_write_quorum_errs(&errs, BUCKET_OP_IGNORED_ERRS, write_quorum) {
-            return Err(err.into());
-        }
-
-        let mut counts: HashMap<String, (usize, BucketInfo)> = HashMap::new();
-        for volumes in infos.into_iter().flatten() {
-            for volume in volumes {
-                if is_reserved_or_invalid_bucket(&volume.name, false) {
-                    continue;
-                }
-
-                let entry = counts.entry(volume.name.clone()).or_insert((
-                    0,
-                    BucketInfo {
-                        name: volume.name.clone(),
-                        created: volume.created,
-                        ..Default::default()
-                    },
-                ));
-                entry.0 += 1;
-            }
-        }
-
-        let mut buckets = counts
-            .into_values()
-            .filter_map(|(count, bucket)| (count >= write_quorum).then_some(bucket))
-            .collect::<Vec<_>>();
-        buckets.sort_by(|left, right| left.name.cmp(&right.name));
-        Ok(buckets)
-    }
-
-    #[tracing::instrument(skip(self))]
-    async fn delete_bucket(&self, bucket: &str, _opts: &DeleteBucketOptions) -> Result<()> {
-        let disks = self.disk_inventory().await;
-        let write_quorum = (disks.len() / 2) + 1;
-
-        let mut futures = Vec::with_capacity(disks.len());
-        for disk in disks.iter().cloned() {
-            let bucket = bucket.to_string();
-            futures.push(async move {
-                match disk {
-                    Some(disk) => disk.delete_volume(&bucket).await,
-                    None => Err(DiskError::DiskNotFound),
-                }
-            });
-        }
-
-        let results = join_all(futures).await;
-        let mut errs = Vec::with_capacity(results.len());
-        let mut recreate = false;
-        for result in results {
-            match result {
-                Ok(()) => errs.push(None),
-                Err(err) => {
-                    if matches!(err, DiskError::VolumeNotEmpty) {
-                        recreate = true;
-                    }
-                    errs.push(Some(err));
-                }
-            }
-        }
-
-        if recreate {
-            for (index, err) in errs.iter().enumerate() {
-                if err.is_none()
-                    && let Some(Some(disk)) = disks.get(index)
-                {
-                    let _ = disk.make_volume(bucket).await;
-                }
-            }
-            return Err(Error::VolumeNotEmpty);
-        }
-
-        if let Some(err) = reduce_write_quorum_errs(&errs, BUCKET_OP_IGNORED_ERRS, write_quorum) {
-            return Err(err.into());
-        }
-
-        Ok(())
-    }
-}
-
 fn check_object_lock_retention_update(bucket: &str, object: &str, obj_info: &ObjectInfo, opts: &ObjectOptions) -> Result<()> {
     if let Some(retention) = &opts.object_lock_retention
         && check_retention_for_modification(
@@ -5004,67 +4796,6 @@ impl SetDisks {
         }
 
         resolve_tiered_decommission_write_quorum_result(&errs, write_quorum, bucket, object)
-    }
-}
-
-#[async_trait::async_trait]
-impl crate::storage_api_contracts::list::ListOperations for SetDisks {
-    type Error = Error;
-    type ListObjectsV2Info = ListObjectsV2Info;
-    type ListObjectVersionsInfo = ListObjectVersionsInfo;
-    type ObjectInfoOrErr = ObjectInfoOrErr;
-    type WalkOptions = WalkOptions;
-    type WalkCancellation = CancellationToken;
-    type WalkResultSender = Sender<ObjectInfoOrErr>;
-
-    #[tracing::instrument(skip(self))]
-    async fn list_objects_v2(
-        self: Arc<Self>,
-        bucket: &str,
-        prefix: &str,
-        continuation_token: Option<String>,
-        delimiter: Option<String>,
-        max_keys: i32,
-        fetch_owner: bool,
-        start_after: Option<String>,
-        incl_deleted: bool,
-    ) -> Result<ListObjectsV2Info> {
-        self.inner_list_objects_v2(
-            bucket,
-            prefix,
-            continuation_token,
-            delimiter,
-            max_keys,
-            fetch_owner,
-            start_after,
-            incl_deleted,
-        )
-        .await
-    }
-
-    #[tracing::instrument(skip(self))]
-    async fn list_object_versions(
-        self: Arc<Self>,
-        bucket: &str,
-        prefix: &str,
-        marker: Option<String>,
-        version_marker: Option<String>,
-        delimiter: Option<String>,
-        max_keys: i32,
-    ) -> Result<ListObjectVersionsInfo> {
-        self.inner_list_object_versions(bucket, prefix, marker, version_marker, delimiter, max_keys)
-            .await
-    }
-
-    async fn walk(
-        self: Arc<Self>,
-        rx: CancellationToken,
-        bucket: &str,
-        prefix: &str,
-        result: Sender<ObjectInfoOrErr>,
-        opts: WalkOptions,
-    ) -> Result<()> {
-        self.walk_internal(rx, bucket, prefix, result, opts).await
     }
 }
